+++ src/main/java/org/opentripplanner/gtfs/mapping/ServiceCalendarDateMapper.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.gtfs.mapping;
 
-import org.opentripplanner.model.ServiceCalendarDate;
+import org.opentripplanner.model.calendar.ServiceCalendarDate;
 import org.opentripplanner.util.MapUtils;
 
 import java.util.Collection;
