+++ src/test/java/org/opentripplanner/model/impl/OtpTransitServiceBuilderTest.java
@@ -11,8 +11,8 @@ import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Frequency;
 import org.opentripplanner.model.TransitEntity;
 import org.opentripplanner.model.Route;
-import org.opentripplanner.model.ServiceCalendar;
-import org.opentripplanner.model.ServiceCalendarDate;
+import org.opentripplanner.model.calendar.ServiceCalendar;
+import org.opentripplanner.model.calendar.ServiceCalendarDate;
 import org.opentripplanner.model.ShapePoint;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripStopTimes;
